refactor(ci): replace standard-version with commitizen and workflow
Due to unnessecary files, standard-version got replaced with commitizen. This can handle everything
automatically with the added workflow (which needs to be enabled). This also means, no nodejs/npm is
more required. But we recommend to add commitizen-cli (cz) as a global package to create easy commit
messages.

BREAKING CHANGE: Replaced standard-version with commitizen